[774] NodeName proper comparison
