+++ bob-common/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod bob_client;
 pub mod configs;
+mod name_types; // Private module. Inner types should be re-exported where they are neede
 pub mod data;
 pub mod error;
 pub mod mapper;
